+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     - `AWSIoTDataPublishRequest`
     - `AWSIoTDataUpdateThingShadowRequest`
     - `AWSIoTDataUpdateThingShadowResponse`
+  - Amazon Kinesis Video Streams ([PR #3161](https://github.com/aws-amplify/aws-sdk-ios/pull/3161))
+  - Amazon Kinesis Video Streams Archived Media ([PR #3161](https://github.com/aws-amplify/aws-sdk-ios/pull/3161))
   - Amazon Lambda ([PR #3154](https://github.com/aws-amplify/aws-sdk-ios/pull/3154)). Note that the following base request and response objects that include untyped (i.e., `id`) properties do not support `NSSecureCoding`. To support `NSSecureCoding` for those types, create a subclass of the base type, and override the appropriate `initWithCoder:` methods to provide a type-safe unarchiving method:
     - `AWSLambdaInvocationRequest`
     - `AWSLambdaInvocationResponse`